security: implement PRD 7.2 security fixes
- Rate limiting on auth routes (Flask-Limiter)
- Validate next parameter to prevent open redirect
- Force SECRET_KEY in production, use _get_secret_key() helper
- HMAC-SHA256 for webhook signatures (fix length extension attack)
- Migrate monetary fields from Float to Numeric(10,2) with Decimal
- HTTPS enforcement via before_request and secure session cookies
- Security headers (CSP, HSTS, X-Frame-Options, etc)

+++ app/models/subscription.py
@@ -34,14 +34,14 @@ class Transaction(db.Model):
     subscription_id = db.Column(db.Integer, db.ForeignKey('subscriptions.id'), nullable=False)
     
     # Valores financeiros
-    amount = db.Column(db.Float, nullable=False)
-    fee = db.Column(db.Float, nullable=False, default=0)  # Compatibilidade
-    net_amount = db.Column(db.Float, nullable=False, default=0)
-    
+    amount = db.Column(db.Numeric(10, 2), nullable=False)
+    fee = db.Column(db.Numeric(10, 2), nullable=False, default=0)  # Compatibilidade
+    net_amount = db.Column(db.Numeric(10, 2), nullable=False, default=0)
+
     # Novas colunas de taxa detalhada
-    fixed_fee = db.Column(db.Float, nullable=False, default=0.99)
-    percentage_fee = db.Column(db.Float, nullable=False, default=0)
-    total_fee = db.Column(db.Float, nullable=False, default=0)
+    fixed_fee = db.Column(db.Numeric(10, 2), nullable=False, default=0.99)
+    percentage_fee = db.Column(db.Numeric(10, 2), nullable=False, default=0)
+    total_fee = db.Column(db.Numeric(10, 2), nullable=False, default=0)
     pix_transaction_id = db.Column(db.String(100))
     
     # Status e método
